// Admin-Portal-KYB-App/microsites-main/packages/case/src/types/risk.ts
export type RiskStatus =
	| "match"
	| "success"
	| "manually_approved"
	| "verified"
	| "yes"
	| "pending"
	| "active"
	| "waiting_for_prerequisite"
	| "pending_review"
	| "partial_match"
	| "no_match"
	| "failed"
	| "expired"
	| "no_data"
	| "no_input"
	| "canceled"
	| "manually_rejected"
	| "no"
	| "error";